[Cherry-pick] Change escompat.Array 'length' property type from 'number' to 'int'
Description: Change escompat.Array 'length' property type from 'number' to 'int'
Issue: https://gitee.com/openharmony/arkcompiler_runtime_core/issues/ICRDCD

是否涉及非兼容变更:否
TDD自验结果:Pass
XTS自验结果:Pass

+++ frameworks/bridge/arkts_frontend/koala_projects/incremental/runtime/src/states/State.ts
@@ -419,7 +419,7 @@ class ArrayStateImpl<Item> extends StateImpl<Array<Item>> implements ArrayState<
     }
 
     set length(value: number) {
-        this.mutable.length = value
+        this.mutable.length = value.toInt()
     }
 
     at(index: number): Item {
@@ -465,7 +465,7 @@ class ArrayStateImpl<Item> extends StateImpl<Array<Item>> implements ArrayState<
 
     splice(start: number, deleteCount: number | undefined, ...items: Item[]): Array<Item> {
         const array = this.mutable
-        return array.splice(start, deleteCount ?? array.length, ...items)
+        return array.splice(start, deleteCount ?? array.length.toDouble(), ...items)
     }
 
     unshift(...items: Item[]): number {
